simulatore-python/requirements.txt
flask
google-generativeai
pytz